% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-discorr.R
\name{discorr}
\alias{discorr}
\alias{pwc}
\title{Construct Distance and Correlation Matrices}
\usage{
discorr(df, longlat = FALSE, ctype = "single1")

pwc(x1, x2, ctype)
}
\arguments{
\item{df}{a fema-class data.frame.}

\item{longlat}{logical indicating whether coordinates are projected.  Default
is \code{FALSE}.}

\item{ctype}{a string indicating the type of correlation to be caluclated.
Valid options include single1 (cosine similarity for a single slope),
single2 (linear similarity for a single slope),
and multi (square root of the sum of squared differences across
multiple slopes).}

\item{x1}{a vector of slopes associated with observation i.}

\item{x2}{a vector of slopes associated with observation j.}
}
\value{
The function \code{discorr} returns the following list of objects:
\describe{
  \item{coords}{matrix of coordinates.}
  \item{slopes}{vectors of slopes.}
  \item{longlat}{logical indicating whether coordinates are projected.}
  \item{dmat}{the distance matrix.}
  \item{rmat}{the correlation matrix.}
}
}
\description{
\code{discorr} is used to construct the distance and correlation matrices
needed to estimate range, organization, and strength.
}
\details{
\code{discorr} is used to generate distance and correlation matrices
that can then be summarized in terms of range, organization, and strength.
\code{pwc} calculates pairwise correlations using the methods outlined by
Martin, Slez, and Borkenhagen (2016).
}
\references{
Martin, J.L., Slez, A., and Borkenhagen, C.  2016.
"Some Provisional Techniques for Quantifying the Degree of Field Effect
in Social Data."
}

