% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEI_utils.R
\name{get_best_points_cnty_sf}
\alias{get_best_points_cnty_sf}
\title{Get best points with county info}
\usage{
get_best_points_cnty_sf()
}
\value{
simple feature
}
\description{
Get best points with county info
}
