% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEI_utils.R
\name{get_best_points_sf}
\alias{get_best_points_sf}
\title{Get hurricane location history
https://www.ncei.noaa.gov/data/international-best-track-archive-for-climate-stewardship-ibtracs/v04r00/access/best_points/IBTrACS.ALL.list.v04r00.points.zip
as POINT geometry}
\usage{
get_best_points_sf()
}
\value{
a simple feature object
}
\description{
Get hurricane location history
https://www.ncei.noaa.gov/data/international-best-track-archive-for-climate-stewardship-ibtracs/v04r00/access/best_points/IBTrACS.ALL.list.v04r00.points.zip
as POINT geometry
}
