% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEI_utils.R
\name{get_best_tracks_dt}
\alias{get_best_tracks_dt}
\title{Get hurricane location history
https://www.ncei.noaa.gov/data/international-best-track-archive-for-climate-stewardship-ibtracs/v04r00/access/csv/ibtracs.ALL.list.v04r00.csv
as POINT geometry}
\usage{
get_best_tracks_dt()
}
\value{
a data.table with units
}
\description{
Get hurricane location history
https://www.ncei.noaa.gov/data/international-best-track-archive-for-climate-stewardship-ibtracs/v04r00/access/csv/ibtracs.ALL.list.v04r00.csv
as POINT geometry
}
