% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCEI_utils.R
\name{get_best_tracks_sf}
\alias{get_best_tracks_sf}
\title{get best tracks
Get hurricane location history
https://www.ncei.noaa.gov/data/international-best-track-archive-for-climate-stewardship-ibtracs/v04r00/access/best_points/IBTrACS.ALL.list.v04r00.lines.zip
as LINES geometry (one feature per hurricane and 6 hours interval)}
\usage{
get_best_tracks_sf()
}
\value{
a simple feature object
}
\description{
get best tracks
Get hurricane location history
https://www.ncei.noaa.gov/data/international-best-track-archive-for-climate-stewardship-ibtracs/v04r00/access/best_points/IBTrACS.ALL.list.v04r00.lines.zip
as LINES geometry (one feature per hurricane and 6 hours interval)
}
