% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRI_utils.R
\name{get_NRI_ctys_sf}
\alias{get_NRI_ctys_sf}
\title{NRI Hazard Info by County}
\usage{
get_NRI_ctys_sf(state)
}
\arguments{
\item{state}{character}
}
\value{
sf simple feature
}
\description{
NRI Hazard Info by County
}
