% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRI_utils.R
\name{get_NRI_HazardInfo}
\alias{get_NRI_HazardInfo}
\title{get_NRI_HazardInfo
Get the main hazrds table from FEMA/NRI}
\usage{
get_NRI_HazardInfo()
}
\description{
get_NRI_HazardInfo
Get the main hazrds table from FEMA/NRI
}
