% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{get_tracts_sf}
\alias{get_tracts_sf}
\title{gt census tracts
Because blocks nest in tracts, and tract IDs even form the first part of block
IDs, it's relatively straightforward to generate tract crosswalks from block
crosswalks.}
\usage{
get_tracts_sf(
  stateabbr = NULL,
  county = NULL,
  year = getOption("tigris_year", 2022L)
)
}
\arguments{
\item{year}{int tigris year}
}
\value{
simple feature
}
\description{
gt census tracts
Because blocks nest in tracts, and tract IDs even form the first part of block
IDs, it's relatively straightforward to generate tract crosswalks from block
crosswalks.
}
