% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/6-gwr2fema.R
\name{gwr2fema}
\alias{gwr2fema}
\title{Convert GWR Output to FEMA Data Frame}
\usage{
gwr2fema(x, slopes = NULL, intercept = FALSE)
}
\arguments{
\item{x}{a gwr-class object}

\item{slopes}{a string vector indicating which slopes to retain.  The default
is \code{NULL}, which retains all slopes.}

\item{intercept}{logical indicating whether or not to return the intercept.
The default is \code{FALSE}.}
}
\value{
The function \code{gwr2fema} returns a fema-class data frame.
}
\description{
\code{gwr2fema} converts output produced using \code{gwr} to a data frame
that can be read by \code{discorr}.
}

