% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\name{get_cty_cbsa_NRI_sf}
\alias{get_cty_cbsa_NRI_sf}
\title{get NRI Hazard Info by County with CBSA}
\usage{
get_cty_cbsa_NRI_sf()
}
\value{
simple feature
}
\description{
get NRI Hazard Info by County with CBSA
}
