% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracts.R
\name{get_NRI_tracts_sf}
\alias{get_NRI_tracts_sf}
\title{NRI Hazard Info by Census Tabulation Block}
\usage{
get_NRI_tracts_sf(state)
}
\arguments{
\item{state}{character}
}
\value{
sf simple feature
}
\description{
NRI Hazard Info by Census Tabulation Block
}
